# Serverless-Finance/lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetterCapitalInvestment API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  message: string;
}