# Serverless-Finance/lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetterCapitalInvestment API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  email: string;
  fullName: string;
  tier: string;
  theme: string;
  biometricEnabled: boolean;
  liquidity: number;
  /** @nullable */
  isAdmin?: boolean | null;
  /** @nullable */
  emailVerified?: boolean | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  bankAccountName?: string | null;
  /** @nullable */
  cryptoWithdrawAddress?: string | null;
  /** @nullable */
  cryptoWithdrawNetwork?: string | null;
}