# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BetterCapitalInvestment API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface UserProfile {
  id: number;
  email: string;
  fullName: string;
  tier: string;
  theme: string;
  biometricEnabled: boolean;
  liquidity: number;
  /** @nullable */
  isAdmin?: boolean | null;
  /** @nullable */
  emailVerified?: boolean | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  bankAccountName?: string | null;
  /** @nullable */
  cryptoWithdrawAddress?: string | null;
  /** @nullable */
  cryptoWithdrawNetwork?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface SignupInput {
  email: string;
  password: string;
  fullName: string;
}

export interface Investment {
  id: string;
  sectorId: string;
  sectorTitle: string;
  amount: number;
  startDateStamp: string;
  daysActive: number;
  dailyRate: number;
  accruedYield: number;
  tierName: string;
  status: string;
}

export interface InvestmentInput {
  sectorId: string;
  sectorTitle: string;
  amount: number;
  dailyRate: number;
  tierName: string;
}

export interface InvestmentUpdate {
  status?: string;
  accruedYield?: number;
  daysActive?: number;
}

export interface Transaction {
  id: string;
  type: string;
  fund: string;
  date: string;
  amount: number;
}

export interface TransactionInput {
  type: string;
  fund: string;
  amount: number;
}

export interface Notification {
  id: string;
  title: string;
  message: string;
  timestamp: string;
  read: boolean;
  type: string;
}

export interface PortfolioSummary {
  totalWealth: number;
  activePrincipal: number;
  totalROIReceived: number;
  liquidity: number;
  activeInvestmentCount: number;
}

export interface LiquidityResponse {
  liquidity: number;
}

export interface LiquidityUpdate {
  delta: number;
}

export interface ProfileUpdate {
  fullName?: string;
  theme?: string;
  biometricEnabled?: boolean;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountName?: string;
  cryptoWithdrawAddress?: string;
  cryptoWithdrawNetwork?: string;
}