# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BetterCapitalInvestment API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    UserProfile:
      type: object
      required: [id, email, fullName, tier, theme, biometricEnabled, liquidity]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        tier:
          type: string
        theme:
          type: string
        biometricEnabled:
          type: boolean
        liquidity:
          type: number
        isAdmin:
          type: ["boolean", "null"]
        emailVerified:
          type: ["boolean", "null"]
        avatarUrl:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccountNumber:
          type: ["string", "null"]
        bankAccountName:
          type: ["string", "null"]
        cryptoWithdrawAddress:
          type: ["string", "null"]
        cryptoWithdrawNetwork:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    SignupInput:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
...